# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Always_date_api documentation",
        "description": "Always date documentation for back-end",
        "contact": {
            "name": "FlokY#0101"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Iegūst lietotāja datus",
                "description": "Iegūst autorizētā",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Rediģē lietotāja datus",
                "description": "Rediģē autorizētā lietotāja datus",
                "operationId": "updateMe",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/users{filter}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Iegūst lietotājus",
                "description": "Iegūst lietotājus pēc filtrācijas. Filtrācijas piemēri:<br/>\n     *          Kārtošana - ?sort[]=name&sort[]=desc/asc;<br/>\n     *          Iegūt pēc filtrācijas - ?name=Alberts<br/>\n     *          Iegūt starp skaitļiem(getBetween) - ?age[]=1&age[]=11;<br/>\n     *          Kombinēti(combined) - ?age[]=1&age[]=11&name=Alberts&sort[]=name&sort[]=asc;",
                "operationId": "getUsersByFilter",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "path",
                        "description": "Filtrācija un kārtošana izmantojot augstāk dotos piemērus.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/rate": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Lietotāju vērtēšana",
                "description": "Lietotāju vērtēšana, kur kāds lietotājs, novērtē kādu citu lietoāju",
                "operationId": "postUserRate",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "Lietotāja id, kurš tiek vēretēts",
                                        "format": "integer"
                                    },
                                    "rating": {
                                        "description": "Vērtējums ar skaitli",
                                        "format": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Jūs jau esat novērtējuši."
                    },
                    "401": {
                        "description": "Nevar novērtēt pats sevi."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/rated_user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Iegūst visus kas novērtēja autorizēto lietotāju",
                "description": "Iegūst visus lietotājus, kuri ir novērtējuši autorizētā lietotāja profilu.",
                "operationId": "getRatedUser",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user_rated": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Iegūst lietotājusi, kuri tika novērtēti ar autorizēto lietotāju",
                "description": "Iegūst visus lietotājus, kur autorizētais lietotājs ir novērtējis, kādu citu lietotāju.",
                "operationId": "getUserRated",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserResource": {
                "title": "User resource",
                "description": "User resource",
                "xml": {
                    "name": "UserResource"
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}